multiple improvements
added support for using non server requiring commands directly in the bots DM.
refreshes a servers users whenever a new one is registered or unlinked.
due to the changes to how a server refreshes the interval in which servers are being automatically refreshed have been increased to 30 min.

+++ guildmanager/client.py
@@ -10,6 +10,7 @@ from .manager import Manager
 from . import exceptions
 from . import PATH
 import discord
+from discord import Message, Member, Role
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 
 # TODO tell ppl not to start command in the private mssage unless told to do so
@@ -21,13 +22,13 @@ class Client(discord.Client):
     def __init__(self) -> None:
         super(Client, self).__init__()
         self.commands = {
-            re.compile(r"^/install$"):    self.install,
-            re.compile(r"^/register"):    self.register,
-            re.compile(r"^/uninstall"):   self.uninstall,
-            re.compile(r"^/unregister$"): self.unregister,
-            re.compile(r"^/refresh$"):    self.refresh,
-            re.compile(r"^/unlink$"):     self.unlink,
-            re.compile(r"^/help$"):       self.help}
+            re.compile(r"^/gm install$"):    self.install,
+            re.compile(r"^/gm register"):    self.register,
+            re.compile(r"^/gm uninstall"):   self.uninstall,
+            re.compile(r"^/gm unregister$"): self.unregister,
+            re.compile(r"^/gm refresh$"):    self.refresh,
+            re.compile(r"^/gm unlink$"):     self.unlink,
+            re.compile(r"^/gm help$"):       self.help}
 
         self.check_reply = {
             "api": re.compile(
@@ -41,7 +42,7 @@ class Client(discord.Client):
 
         self.db = "db.db"
 
-    async def check_commands(self, message: discord.Message) -> None:
+    async def check_commands(self, message: Message) -> None:
         """
         checks a message for commands, if found call its method
 
@@ -49,11 +50,14 @@ class Client(discord.Client):
         :return: None
         """
         for command, func in self.commands.items():
-            if command.fullmatch(message.content):
-                await message.delete()
+            if command.fullmatch(message.content.lower()):
+                try:
+                    await message.delete()
+                except (discord.errors.NotFound, discord.errors.Forbidden):
+                    pass
                 await func(message)
 
-    async def _ask_for(self, key: str, message: discord.Message, question: str) -> discord.Message:
+    async def _ask_for(self, key: str, message: Message, question: str) -> Message:
         """
         ask the user for something
 
@@ -69,7 +73,7 @@ class Client(discord.Client):
             await message.author.dm_channel.send(question)
 
         try:
-            reply: discord.Message = await self.wait_for(
+            reply = await self.wait_for(
                 event="message",
                 check=lambda m: m.author == message.author,
                 timeout=300)
@@ -112,12 +116,9 @@ class Client(discord.Client):
                     if not message.author.dm_channel:
                         await message.author.create_dm()
 
-                    reply: discord.Message = await self._ask_for(
-                        "2", message, mes.install.introduction)
-
+                    reply = await self._ask_for("2", message, mes.install.introduction)
                     if reply.content == "1":
-                        reply: discord.Message = await self._ask_for(
-                            "api", message, mes.install.provide_api)
+                        reply = await self._ask_for("api", message, mes.install.provide_api)
                         api = reply.content
                         async with Requester(api) as session:
                             guilds = await session.get_guilds()
@@ -125,11 +126,8 @@ class Client(discord.Client):
                                 "[placeholder]", "\n".join(
                                     [f"{index+1}: {guild['name']} [{guild['tag']}]"
                                      for index, guild in enumerate(guilds)])))
-
-                            reply: discord.Message = await self._ask_for(
-                                str(len(guilds)), message, question)
+                            reply = await self._ask_for(str(len(guilds)), message, question)
                             guild = guilds[int(reply.content) - 1]
-
                             owner_gw2_name = await session.get_gw2_username()
 
                         try:
@@ -147,9 +145,11 @@ class Client(discord.Client):
                                     try:
                                         await manager.add_user(owner_discord_id, owner_gw2_name, api)
                                     except exceptions.UserAlreadyExcists:
+                                        # TODO notefy the user that it already excists
                                         pass
                                     await manager.add_user_server(owner_discord_id, server_id)
                                     await message.author.dm_channel.send(mes.install.registered)
+                                    await self.refresh(message)
                                 except exceptions.UsersServersLinkAlreadyExcists:
                                     message.author.dm_channel.send(mes.install.already_excists)
                             else:
@@ -200,8 +200,7 @@ class Client(discord.Client):
                     )
 
     @staticmethod
-    async def _register_with_excisting_user(
-            message: discord.Message, manager: Manager, user: dict) -> None:
+    async def _register_with_excisting_user(message: Message, manager: Manager, user: dict) -> None:
         """
         tries to link an excisting user to a server
 
@@ -232,7 +231,7 @@ class Client(discord.Client):
             await message.author.dm_channel.send(mes.register.server_not_installed)
 
     @staticmethod
-    async def _remove_roles(member: discord.Member, *roles: discord.Role) -> None:
+    async def _remove_roles(member: Member, *roles: Role) -> None:
         """
         checks if a user have a role, if the user have it it gets removed
 
@@ -243,7 +242,7 @@ class Client(discord.Client):
         await member.remove_roles(*[role for role in roles if role in member.roles])
 
     @staticmethod
-    async def _add_roles(member: discord.Member, *roles: discord.Member) -> None:
+    async def _add_roles(member: Member, *roles: Member) -> None:
         """
         checs if a user have a role, if not add it
 
@@ -253,7 +252,7 @@ class Client(discord.Client):
         """
         await member.add_roles(*[role for role in roles if role not in member.roles])
 
-    async def register(self, message: discord.Message) -> None:
+    async def register(self, message: Message) -> None:
         """
         registers a user to the bot
 
@@ -263,24 +262,24 @@ class Client(discord.Client):
         :return: None
         """
         if "guild" in dir(message.channel):
+            # await message.delete()
             if message.channel.guild:
                 await self._ensure_dm_excists(message)
                 async with Manager(self.db) as manager:
                     user = await manager.get_user(message.author.id)
                     if user:
                         await self._register_with_excisting_user(message, manager, user)
+                        await self.refresh(message)
                     else:
-                        reply: discord.Message = await self._ask_for(
-                            "3", message, mes.register.introduction)
-
+                        reply = await self._ask_for("3", message, mes.register.introduction)
                         if reply.content == "1":
-                            reply: discord.Message = await self._ask_for(
-                                "api", message, mes.register.provide_api)
+                            reply = await self._ask_for("api", message, mes.register.provide_api)
                             async with Requester(reply.content) as session:
                                 gw2_name = await session.get_gw2_username()
                             await manager.add_user(message.author.id, gw2_name, reply.content)
                             user = await manager.get_user(message.author.id)
                             await self._register_with_excisting_user(message, manager, user)
+                            await self.refresh(message)
 
                         elif reply.content == "2":
                             await message.author.dm_channel.send(mes.register.api_help_1)
@@ -288,16 +287,16 @@ class Client(discord.Client):
                             await message.author.dm_channel.send(mes.register.api_help_2)
                             await asyncio.sleep(2)
                             await message.author.dm_channel.send(mes.register.api_help_3)
-                            reply: discord.Message = await self._ask_for(
-                                "2", message, mes.register.api_help_4)
+                            reply = await self._ask_for("2", message, mes.register.api_help_4)
 
                             if reply.content == "1":
                                 await self.register(message)
+                                await self.refresh(message)
 
                         else:
                             await message.author.dm_channel.send(mes.register.offended)
 
-    async def unregister(self, message: discord.Message) -> None:
+    async def unregister(self, message: Message) -> None:
         """
         unregisters a user from the bot
 
@@ -306,26 +305,24 @@ class Client(discord.Client):
         :param message: original message that invoked the remove command
         :return: None
         """
-        if "guild" in dir(message.channel):
-            if message.channel.guild:
-                await self._ensure_dm_excists(message)
-                async with Manager(self.db) as manager:
-                    user = await manager.get_user(message.author.id)
-                    if user:
-                        reply: discord.Message = await self._ask_for(
-                            "2", message, mes.unregister.are_you_sure)
-                        if reply.content == "1":
-                            try:
-                                await manager.remove_user(message.author.id)
-                                await message.author.dm_channel.send(mes.unregister.success)
-                            except exceptions.MissingUser:
-                                await message.author.dm_channel.send(mes.unregister.not_registered)
-                        else:
-                            await message.author.dm_channel.send(mes.unregister.regret)
-                    else:
-                        await message.author.dm_channel.send(mes.unregister.not_registered)
+        await self._ensure_dm_excists(message)
 
-    async def uninstall(self, message: discord.Message) -> None:
+        async with Manager(self.db) as manager:
+            user = await manager.get_user(message.author.id)
+            if user:
+                reply = await self._ask_for("2", message, mes.unregister.are_you_sure)
+                if reply.content == "1":
+                    try:
+                        await manager.remove_user(message.author.id)
+                        await message.author.dm_channel.send(mes.unregister.success)
+                    except exceptions.MissingUser:
+                        await message.author.dm_channel.send(mes.unregister.not_registered)
+                else:
+                    await message.author.dm_channel.send(mes.unregister.regret)
+            else:
+                await message.author.dm_channel.send(mes.unregister.not_registered)
+
+    async def uninstall(self, message: Message) -> None:
         """
         uninstalls the bot from the discord server/gw2 guild
 
@@ -335,14 +332,14 @@ class Client(discord.Client):
         :return: None
         """
         if "guild" in dir(message.channel):
+            # await message.delete()
             if message.channel.guild:
                 await self._ensure_dm_excists(message)
                 async with Manager(self.db) as manager:
                     server = await manager.get_server(message.channel.guild.id)
                     # TODO change to server excists
                     if server:
-                        reply: discord.Message = await self._ask_for(
-                            "2", message, mes.uninstall.are_you_sure)
+                        reply = await self._ask_for("2", message, mes.uninstall.are_you_sure)
                         if reply.content == "1":
                             try:
                                 await manager.remove_server(message.channel.guild.id)
@@ -354,7 +351,7 @@ class Client(discord.Client):
                     else:
                         await message.author.dm_channel.send(mes.uninstall.not_installed)
 
-    async def unlink(self, message: discord.Message) -> None:
+    async def unlink(self, message: Message) -> None:
         """
         unlinks a user from a server
 
@@ -364,27 +361,28 @@ class Client(discord.Client):
         :param message: original message that invoked the /unlink command
         :return: None
         """
-        if message.channel.guild:
-            await self._ensure_dm_excists(message)
-            async with Manager(self.db) as manager:
-                link = await manager.get_users_servers(message.author.id, message.channel.guild.id)
-                # TODO implement Manager.link_excists()
-                servers = [server["server_id"] for server in link]
-                if message.channel.guild.id in servers:
-                    reply: discord.Message = await self._ask_for(
-                        "2", message, mes.unlink.are_you_sure)
-                    if reply.content == "1":
-                        try:
-                            await manager.unlink_user(message.author.id, message.guild.id)
-                            await message.author.dm_channel.send(mes.unlink.success)
-                        except exceptions.MissingUsersServer:
-                            await message.author.dm_channel.send(mes.unlink.not_linked)
+        if "guild" in dir(message.channel):
+            if message.channel.guild:
+                await self._ensure_dm_excists(message)
+                async with Manager(self.db) as manager:
+                    link = await manager.get_users_servers(message.author.id, message.channel.guild.id)
+                    # TODO implement Manager.link_excists()
+                    servers = [server["server_id"] for server in link]
+                    if message.channel.guild.id in servers:
+                        reply = await self._ask_for("2", message, mes.unlink.are_you_sure)
+                        if reply.content == "1":
+                            try:
+                                await manager.unlink_user(message.author.id, message.guild.id)
+                                await message.author.dm_channel.send(mes.unlink.success)
+                                await self.refresh(message)
+                            except exceptions.MissingUsersServer:
+                                await message.author.dm_channel.send(mes.unlink.not_linked)
+                        else:
+                            await message.author.dm_channel.send(mes.unlink.regret)
                     else:
-                        await message.author.dm_channel.send(mes.unlink.regret)
-                else:
-                    await message.author.dm_channel.send(mes.unlink.not_linked)
+                        await message.author.dm_channel.send(mes.unlink.not_linked)
 
-    async def help(self, message: discord.Message) -> None:
+    async def help(self, message: Message) -> None:
         """
         sends a help message to the user who invoked the command
 
@@ -394,7 +392,7 @@ class Client(discord.Client):
         await self._ensure_dm_excists(message)
         await message.author.dm_channel.send(mes.help.commands)
 
-    async def refresh(self, message: discord.Message) -> None:
+    async def refresh(self, message: Message) -> None:
         """
         user prompted refresh command, refreshes a single server
 
@@ -404,14 +402,19 @@ class Client(discord.Client):
         :return: None
         """
         if "guild" in dir(message.channel):
+            # await message.delete()
             async with Manager(self.db) as manager:
                 server_data = await manager.get_server(message.guild.id)
-                server = self.get_guild(server_data["server_id"])
-            if server:
-                await self._refresh_server(server, server_data, manager)
-            else:
-                # TODO error message here
-                pass
+                if server_data:
+                    server = self.get_guild(server_data["server_id"])
+                    if server:
+                        await self._refresh_server(server, server_data, manager)
+                        await message.author.dm_channel.send(mes.refresh.success)
+                    else:
+                        # TODO error message here
+                        pass
+                else:
+                    await message.author.dm_channel.send(mes.refresh.missing_server)
 
     async def _refresh_all(self) -> None:
         """
@@ -459,7 +462,7 @@ class Client(discord.Client):
                     else:
                         await self._remove_roles(server_member, role)
 
-    async def rename(self, message: discord.message) -> None:
+    async def rename(self, message: Message) -> None:
         """
         merges roles that have been renamed from ingame
 
@@ -483,10 +486,10 @@ class Client(discord.Client):
 
         await self._refresh_all()
         scheduler = AsyncIOScheduler()
-        scheduler.add_job(self._refresh_all, "interval", minutes=10)
+        scheduler.add_job(self._refresh_all, "interval", minutes=30)
         scheduler.start()
 
-    async def on_message(self, message: discord.Message) -> None:
+    async def on_message(self, message: Message) -> None:
         """
         command handle
 
@@ -507,6 +510,9 @@ class Client(discord.Client):
             file = json.load(f)
         token = file["client_token"]
         self.run(token)
+
+    def __repr__(self):
+        return f"{self.__class__.__name__} from file {__file__}"
 
 
 if __name__ == '__main__':
